move helpers to corresponding modules and delete locatorUtils
