+++ packages/visual-editor/src/components/locator/Locator.tsx
@@ -1,12 +1,6 @@
 import { FieldLabel, setDeep } from "@puckeditor/core";
-import React from "react";
-import {
-  type MultiSelectorOption,
-  type MultiSelectorValue,
-} from "../../fields/MultiSelectorField.tsx";
+import { type MultiSelectorValue } from "../../fields/MultiSelectorField.tsx";
 import { YextAutoField } from "../../fields/YextAutoField.tsx";
-import { useDocument } from "../../hooks/useDocument.tsx";
-import { useTemplateMetadata } from "../../internal/hooks/useMessageReceivers.ts";
 import { TranslatableString } from "../../types/types.ts";
 import { msg, pt } from "../../utils/i18n/platform.ts";
 import { TranslatableAssetImage } from "../../types/images.ts";
@@ -28,409 +22,24 @@ import {
   YextFields,
 } from "../../fields/fields.ts";
 import { ImageStylingFields } from "../contentBlocks/image/styling.ts";
+import { getFacetFieldOptions } from "./Filters.tsx";
 import {
   DEFAULT_LOCATOR_RESULT_CARD_PROPS,
   DistanceDisplayOption,
-  LocatorResultCardFields,
   LocatorResultCardProps,
 } from "./LocatorResultCard.tsx";
+import { ResultCardPropsField } from "./Results.tsx";
 import { LocatorWrapper } from "./LocatorWrapper.tsx";
 import {
-  DEFAULT_DISTANCE_DISPLAY,
   DEFAULT_LOCATION_STYLE,
   DEFAULT_MAKI_ICON_NAME,
   DEFAULT_PIN_ICON_WIDTH,
-  DEFAULT_TITLE,
-  getLocatorConfigFromPageSet,
   LOCATOR_PIN_ICON_FIELD,
-  makiIconOptions,
   MAX_PIN_ICON_WIDTH,
-} from "./locatorUtils.ts";
-
-const BOOLEAN_SUPPORTED_FIELDS = [
-  "primaryHeading",
-  "secondaryHeading",
-  "tertiaryHeading",
-] as const;
-
-const ResultCardPropsField = ({
-  value,
-  onChange,
-}: {
-  value?: LocatorResultCardProps;
-  onChange: (value: LocatorResultCardProps) => void;
-}) => {
-  const streamDocument = useDocument();
-  const templateMetadata = useTemplateMetadata();
-  const entityTypeSourceMap = getLocatorEntityTypeSourceMap();
-  const entityTypeScopes = React.useMemo(() => {
-    const locatorConfig = getLocatorConfigFromPageSet(streamDocument?._pageset);
-    return locatorConfig.entityTypeScope ?? [];
-  }, [streamDocument]);
-
-  /**
-   * Builds the field schema for the result card editor, including:
-   * - Conditionally removing the primary CTA section when entity scope is not attached to a page set.
-   * - Toggling constant value vs. field selector visibility per section.
-   */
-  const resultCardFields = React.useMemo(() => {
-    if (!value?.entityType) {
-      return LocatorResultCardFields;
-    }
-    let fields = LocatorResultCardFields;
-    const entityTypeHasSourcePageSet = !!entityTypeSourceMap[value.entityType];
-    const scopeExistsForEntityType =
-      entityTypeScopes.find(
-        (scope) => scope.entityType === value.entityType
-      ) !== undefined;
-
-    fields = setDeep(
-      fields,
-      `objectFields.primaryCTA.objectFields.link.visible`,
-      !entityTypeHasSourcePageSet && scopeExistsForEntityType
-    );
-
-    // For each section, show either the field selector or the constant value editor.
-    BOOLEAN_SUPPORTED_FIELDS.forEach((key) => {
-      const headingConfig = value[key];
-      const constantValueEnabled = headingConfig?.constantValueEnabled ?? false;
-      const field = headingConfig?.field;
-      const fieldTypeId = field
-        ? templateMetadata?.locatorDisplayFields?.[field]?.field_type_id
-        : undefined;
-      const booleanFieldSelected =
-        !constantValueEnabled && fieldTypeId === "type.boolean";
-
-      fields = setDeep(
-        fields,
-        `objectFields.${key}.objectFields.field.visible`,
-        !constantValueEnabled
-      );
-      fields = setDeep(
-        fields,
-        `objectFields.${key}.objectFields.constantValue.visible`,
-        constantValueEnabled
-      );
-      fields = setDeep(
-        fields,
-        `objectFields.${key}.objectFields.trueDisplayText.visible`,
-        !constantValueEnabled && booleanFieldSelected
-      );
-      fields = setDeep(
-        fields,
-        `objectFields.${key}.objectFields.falseDisplayText.visible`,
-        booleanFieldSelected
-      );
-    });
-
-    const imageConstantValueEnabled =
-      value.image?.constantValueEnabled ?? false;
-    fields = setDeep(
-      fields,
-      "objectFields.image.objectFields.field.visible",
-      !imageConstantValueEnabled
-    );
-    fields = setDeep(
-      fields,
-      "objectFields.image.objectFields.constantValue.visible",
-      imageConstantValueEnabled
-    );
-
-    return fields;
-  }, [entityTypeSourceMap, entityTypeScopes, templateMetadata, value]);
-
-  return (
-    <YextAutoField
-      field={resultCardFields}
-      value={value ?? DEFAULT_LOCATOR_RESULT_CARD_PROPS}
-      onChange={onChange}
-    />
-  );
-};
-
-function getFacetFieldOptions(
-  entityTypes: LocatorEntityType[]
-): MultiSelectorOption<string>[] {
-  const facetFields: MultiSelectorOption<string>[] = [];
-  const addedValues: Set<string> = new Set<string>();
-  entityTypes.forEach((entityType) =>
-    getFacetFieldOptionsForEntityType(entityType).forEach((option) => {
-      if (option?.value && !addedValues.has(option.value)) {
-        facetFields.push(option);
-        addedValues.add(option.value);
-      }
-    })
-  );
-  return facetFields.sort((a, b) => a.label.localeCompare(b.label));
-}
-
-function getFacetFieldOptionsForEntityType(
-  entityType: LocatorEntityType
-): MultiSelectorOption<string>[] {
-  let filterOptions: MultiSelectorOption<string>[] = [
-    {
-      label: msg("fields.options.facets.city", "City"),
-      value: "address.city",
-    },
-    {
-      label: msg("fields.options.facets.postalCode", "Postal Code"),
-      value: "address.postalCode",
-    },
-    {
-      label: msg("fields.options.facets.region", "Region"),
-      value: "address.region",
-    },
-    {
-      label: msg("fields.options.facets.brandName", "Brand Name"),
-      value: "brandReference.name",
-    },
-  ];
-  switch (entityType) {
-    case "location":
-      filterOptions = filterOptions.concat(
-        {
-          label: msg("fields.options.facets.associations", "Associations"),
-          value: "associations",
-        },
-        {
-          label: msg("fields.options.facets.brands", "Brands"),
-          value: "brands",
-        },
-        {
-          label: msg("fields.options.facets.keywords", "Keywords"),
-          value: "keywords",
-        },
-        {
-          label: msg("fields.options.facets.languages", "Languages"),
-          value: "languages",
-        },
-        {
-          label: msg("fields.options.facets.paymentOptions", "Payment Options"),
-          value: "paymentOptions",
-        },
-        {
-          label: msg("fields.options.facets.products", "Products"),
-          value: "products",
-        },
-        {
-          label: msg("fields.options.facets.services", "Services"),
-          value: "services",
-        },
-        {
-          label: msg("fields.options.facets.specialties", "Specialties"),
-          value: "specialities",
-        }
-      );
-      break;
-    case "restaurant":
-      filterOptions = filterOptions.concat(
-        {
-          label: msg(
-            "fields.options.facets.acceptsReservations",
-            "Accepts Reservations"
-          ),
-          value: "acceptsReservations",
-        },
-        {
-          label: msg("fields.options.facets.associations", "Associations"),
-          value: "associations",
-        },
-        {
-          label: msg("fields.options.facets.brands", "Brands"),
-          value: "brands",
-        },
-        {
-          label: msg("fields.options.facets.keywords", "Keywords"),
-          value: "keywords",
-        },
-        {
-          label: msg("fields.options.facets.languages", "Languages"),
-          value: "languages",
-        },
-        {
-          label: msg("fields.options.facets.mealsServed", "Meals Served"),
-          value: "mealsServed",
-        },
-        {
-          label: msg("fields.options.facets.neighborhood", "Neighborhood"),
-          value: "neighborhood",
-        },
-        {
-          label: msg("fields.options.facets.paymentOptions", "Payment Options"),
-          value: "paymentOptions",
-        },
-        {
-          label: msg(
-            "fields.options.facets.pickupAndDeliveryServices",
-            "Pickup and Delivery Services"
-          ),
-          value: "pickupAndDeliveryServices",
-        },
-        {
-          label: msg("fields.options.facets.priceRange", "Price Range"),
-          value: "priceRange",
-        },
-        {
-          label: msg("fields.options.facets.services", "Services"),
-          value: "services",
-        },
-        {
-          label: msg("fields.options.facets.specialties", "Specialties"),
-          value: "specialities",
-        }
-      );
-      break;
-    case "healthcareFacility":
-      filterOptions = filterOptions.concat(
-        {
-          label: msg(
-            "fields.options.facets.acceptingNewPatients",
-            "Accepting New Patients"
-          ),
-          value: "acceptingNewPatients",
-        },
-        {
-          label: msg(
-            "fields.options.facets.conditionsTreated",
-            "Conditions Treated"
-          ),
-          value: "conditionsTreated",
-        },
-        {
-          label: msg(
-            "fields.options.facets.insuranceAccepted",
-            "Insurance Accepted"
-          ),
-          value: "insuranceAccepted",
-        },
-        {
-          label: msg("fields.options.facets.paymentOptions", "Payment Options"),
-          value: "paymentOptions",
-        },
-        {
-          label: msg("fields.options.facets.services", "Services"),
-          value: "services",
-        }
-      );
-      break;
-    case "healthcareProfessional":
-      filterOptions = filterOptions.concat(
-        {
-          label: msg(
-            "fields.options.facets.acceptingNewPatients",
-            "Accepting New Patients"
-          ),
-          value: "acceptingNewPatients",
-        },
-        {
-          label: msg(
-            "fields.options.facets.admittingHospitals",
-            "Admitting Hospitals"
-          ),
-          value: "admittingHospitals",
-        },
-        {
-          label: msg("fields.options.facets.brands", "Brands"),
-          value: "brands",
-        },
-        {
-          label: msg("fields.options.facets.certifications", "Certifications"),
-          value: "certifications",
-        },
-        {
-          label: msg(
-            "fields.options.facets.conditionsTreated",
-            "Conditions Treated"
-          ),
-          value: "conditionsTreated",
-        },
-        {
-          label: msg("fields.options.facets.degrees", "Degrees"),
-          value: "degrees",
-        },
-        {
-          label: msg("fields.options.facets.gender", "Gender"),
-          value: "gender",
-        },
-        {
-          label: msg(
-            "fields.options.facets.insuranceAccepted",
-            "Insurance Accepted"
-          ),
-          value: "insuranceAccepted",
-        },
-        {
-          label: msg("fields.options.facets.languages", "Languages"),
-          value: "languages",
-        },
-        {
-          label: msg("fields.options.facets.neighborhood", "Neighborhood"),
-          value: "neighborhood",
-        },
-        {
-          label: msg("fields.options.facets.officeName", "Office Name"),
-          value: "officeName",
-        },
-        {
-          label: msg("fields.options.facets.services", "Services"),
-          value: "services",
-        }
-      );
-      break;
-    case "hotel":
-      filterOptions = filterOptions.concat(
-        { label: msg("fields.options.facets.bar", "Bar"), value: "bar" },
-        {
-          label: msg("fields.options.facets.catsAllowed", "Cats Allowed"),
-          value: "catsAllowed",
-        },
-        {
-          label: msg("fields.options.facets.dogsAllowed", "Dogs Allowed"),
-          value: "dogsAllowed",
-        },
-        {
-          label: msg("fields.options.facets.parking", "Parking"),
-          value: "parking",
-        },
-        { label: msg("fields.options.facets.pools", "Pools"), value: "pools" }
-      );
-      break;
-    case "financialProfessional":
-      filterOptions = filterOptions.concat(
-        {
-          label: msg("fields.options.facets.certifications", "Certifications"),
-          value: "certifications",
-        },
-        {
-          label: msg("fields.options.facets.interests", "Interests"),
-          value: "interests",
-        },
-        {
-          label: msg("fields.options.facets.languages", "Languages"),
-          value: "languages",
-        },
-        {
-          label: msg("fields.options.facets.services", "Services"),
-          value: "services",
-        },
-        {
-          label: msg("fields.options.facets.specialties", "Specialties"),
-          value: "specialties",
-        },
-        {
-          label: msg(
-            "fields.options.facets.yearsOfExperience",
-            "Years of Experience"
-          ),
-          value: "yearsOfExperience",
-        }
-      );
-      break;
-    default:
-      break;
-  }
-  return filterOptions;
-}
+  makiIconOptions,
+} from "./Map.tsx";
+const DEFAULT_TITLE = "Find a Location";
+const DEFAULT_DISTANCE_DISPLAY = "distanceFromUser";
 
 export interface LocatorProps {
   /**
